requirements.txt
streamlit
requests
pdfplumber
python-dotenv
